+++ front/src/features/interview/hooks/useCreateInterview.ts
@@ -1,4 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
+import { useNavigate } from 'react-router-dom'
 import { createInterview, createSession, createQuestions } from '../api/interviewApi'
 import type { CreateInterviewRequest } from '../api/interviewApi'
 import { useInterviewStore } from '../stores/interviewStore'
@@ -8,6 +9,7 @@ import { useToast } from '../../../shared/components/ui/toastStore'
 export const useCreateInterview = () => {
   const { setInterview, setSession, setPhase, setQuestionCount } = useInterviewStore()
   const { toast } = useToast()
+  const navigate = useNavigate()
 
   const { mutate, isPending } = useMutation({
     onMutate: () => {
@@ -24,6 +26,7 @@ export const useCreateInterview = () => {
       setSession(sessionId)
       setQuestionCount(questionCount)
       setPhase('chat')
+      navigate(`/interview/${interviewId}`)
     },
     onError: (error: unknown) => {
       setPhase('setup')
