+++ front/src/shared/components/layout/AppLayout.tsx
@@ -30,7 +30,7 @@ const AppLayout = () => {
             <NavLink
               key={to}
               to={to}
-              end={to === '/'}
+              end={to === '/' || to === '/interview'}
               className={({ isActive }) =>
                 `block px-3 py-2 rounded-md text-sm font-medium transition-colors ${
                   isActive
